`XmlRenderer` moved to `Nodes`.
